requirements.txt
boto==2.39.0
dj-database-url==0.4.0
Django==1.9.4
django-betterforms==1.1.4
django-crispy-forms==1.6.0
django-email-obfuscator==0.1.4
django-filter==0.13.0
django-localflavor==1.2
django-multiupload==0.5
django-registration-redux==1.4
django-storages-redux==1.3.2
geopy==1.11.0
gunicorn==19.4.5
more-itertools==2.2
Pillow==3.1.1
psycopg2==2.6.1
requests==2.9.1
six==1.10.0
South==1.0.2
stripe==1.30.0
